requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch
torchvision
argparse
Pillow
requests
tqdm
open_clip_torch
git+https://github.com/huggingface/transformers.git
git+https://github.com/theovercomer8/BLIP
contexttimer
einops>=0.4.1
fairscale==0.4.4
ftfy
iopath
ipython
omegaconf
opencv-python-headless==4.5.5.64
opendatasets
packaging
pandas
plotly
pre-commit
pycocoevalcap
pycocotools
python-magic
scikit-image
sentencepiece
spacy
streamlit
timm==0.4.12
torch>=1.10.0
torchvision
tqdm
webdataset
wheel
salesforce-lavis --install-option="--no-deps"
